refactored graphData and options manipulations into apexChartHelpers.js
